Breaking change: Remove pageable APIs and cache static data (#98)
Remove the `Pageable` interface that is currently used for commands with
no interesting second pages and are never used by the device in
practice. The `page` argument is removed from `get_available_stations`
and `get_zone_states` commands

Adds a cache for commands that are static data about the device, for use
in future more complex calls that care about the type of the device.

Improves error handling for common error cases.