Add defensive checks against unexpected schedule responses (#305)
Add defensive checks against unexpected schedule responses. I do not yet
know the valid version of responses for this command, so for now it
seems best to just check for out of bounds messages and assume that the
schedules are not supported. This can be revisited as we capture traces
for these devices.

For https://github.com/home-assistant/home-assistant.io/issues/29766